Add files via upload
Three-Card-Monte is a common scam played  on gullible tourists and other easy marks. Three playing cards, one of which is the 'red lady' Queen of Hearts, are put facedown on a cardboard box. The dealer quickly rearranges the cards and then asks the mark to pick the Queen of Hearts. But the dealer can use all sort of tricks to hide the card or otherwise cheat, guaranteeing that the victim never wins. It's also common for the dealer to have shills in the crowd who secretly work with the dealer but pretent to win the game (to make the victim think they too could win) or purposefully lose badly (to make the victim think they could do much better).
This program shows the three cards and then quickly describes a series of swaps. At the end, it clears the screen, and the player must pick a card.